requirements.txt
### === Module: requirements.txt ===
# Backtrader 量化交易回測系統依賴套件

# 核心回測框架
backtrader>=1.9.78.123

# 資料下載
yfinance>=0.2.18

# 快取管理
joblib>=1.3.2

# 資料處理
pandas>=2.0.3
numpy>=1.24.3

# 設定檔處理
PyYAML>=6.0.1

# 日期時間處理
python-dateutil>=2.8.2

# 日誌處理
loguru>=0.7.0

# 進度條顯示
tqdm>=4.65.0

# 數值計算
scipy>=1.11.1

# 圖表繪製 (可選)
matplotlib>=3.7.2